backend/pyproject.toml
[project]
name = "ldaca-backend"
version = "0.1.0"
description = "LDaCA Web App Backend - FastAPI backend for the Language Data Commons of Australia"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi-users[sqlalchemy]>=14.0.1",
    "fastapi[standard]>=0.115.14",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "pandas>=2.3.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "sqlalchemy[asyncio]>=2.0.0",
    "uvicorn>=0.35.0",
    "docframe",
    "docworkspace",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::sqlalchemy.exc.SADeprecationWarning",
    "ignore::sqlalchemy.exc.SAWarning"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "workspace: marks tests related to workspace functionality",
]
